refactor: 채팅 구독 관련 코드 useChatSubscription hook으로 분리
